Use unofficial sqlite, fix warnings & time/rng
CI: include vcpkg-configuration.json. CMake: switch to unofficial-sqlite3 (find_package + link target) and remove legacy SQLite3 checks. Database/MainFrame: suppress unused exception variable warnings in catch blocks. Models: make getCurrentTimestamp thread-safe using localtime_s/localtime_r and add <ctime>. Security: fix uniform_int_distribution index type with static_cast<int>. These changes improve portability, silence warnings, and ensure correct cross-platform behavior.

+++ src/services/SecurityManager.cpp
@@ -313,7 +313,7 @@ std::string SecurityManager::generateSessionToken() const {
     const std::string chars = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
     std::random_device rd;
     std::mt19937 gen(rd());
-    std::uniform_int_distribution<> dis(0, chars.size() - 1);
+    std::uniform_int_distribution<> dis(0, static_cast<int>(chars.size() - 1));
     
     std::string token;
     for (int i = 0; i < 32; ++i) {
